Add function to play a single round of the game
New function takes the player's and computer's choices and compares
them to determine the winner (or a tie).

As the user will be typing in either rock, paper, or scissor, the
function that determines the computer's choice was changed so that it
would return not a number, but the actual words that the player would
type in. To disregard capitalizations, the player's input was
uppercased.