Add function so that the game will play 5 rounds
To easily play a game of rock, paper, and scissors, the playRound
function was called by the playGame function 5 times using a for loop.
After the rounds are finished, the function prints out the winner.

The playRound function was fixed because the human and computer scores
were not updating. Upon closer inspection, the function was returning
before the score incrementation code was reached.