Fix problem with post-game incrementation
The gaming buttons were removed after the game ends to prevent the
player from clicking on them and causing the scores to increment even
after the game is over.

After the game is over, a replay function was created giving the player
the choice to play again or quit. There are still problems with the
response of these buttons when they are pressed: the buttons do not
disappear even after the user presses yes, the no button seems to do
nothing, the play again question continues to be displayed.